e2e-tests/installer/pyproject.toml
# SPDX-FileCopyrightText: 2025 Intel Corporation
#
# SPDX-License-Identifier: Apache-2.0

[tool.poetry]
name = "orch-installer-test"
version = "0.1.0"
description = "Orchestrator Installer Test"
authors = ["Matthew Balvin <matthew.d.balvin@intel.com>", "Abhinav Veerasangappala Shridatta <abhinav.veerasangappala.shridatta@intel.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.8"
pexpect = "^4.9.0"
selenium = "4.21.0"
python-dotenv = "1.0.1"
ptyprocess = "0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
